move genesis processing to separate file & genesis chunk -> genesis batch
